pyproject.toml
[project]
name = "cvin3d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.1",
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pytorch-lightning>=2.5.6",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "torch-geometric>=2.7.0",
]
